# server/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager Rest Api",
        "description": "This is documentation for task manager rest api ",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Samiul Islam",
            "email": "samiulislamw.w.w@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    },
    "tags": [
        {
            "name": "list",
            "description": "list curd api document"
        },
        {
            "name": "user",
            "description": "user api CURD and login "
        },
        {
            "name": "task",
            "description": "task curd api  document"
        }
    ],
    "paths": {
        "/lists/create": {
            "post": {
                "tags": [
                    "list"
                ],
                "summary": "List Create Operation",
                "description": "This is can create a new List",
                "operationId": "createList",
                "requestBody": {
                    "description": "Create List Model Object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ListDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {}
                    }
                },
                "x-codegen-request-body-name": "body"
            }
        },
        "/lists/all": {
            "get": {
                "tags": [
                    "list"
                ],
                "summary": "Get all list",
                "description": "A user can get all List",
                "operationId": "getAllList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/lists/{id}": {
            "get": {
                "tags": [
                    "list"
                ],
                "summary": "Find list by id",
                "description": "A user can find one list by her id",
                "operationId": "findOneList",
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/lists/update/{id}": {
            "patch": {
                "tags": [
                    "list"
                ],
                "summary": "Update single list",
                "discription": "a user can update single list",
                "operationId": "updateList",
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "you can update any of single column data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {}
                }
            }
        },
        "/lists/delete/{id}": {
            "delete": {
                "tags": [
                    "list"
                ],
                "summary": "List delete",
                "discription": "a user can delete spacific list",
                "operationId": "deleteList",
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/lists/{listId}/task/create": {
            "post": {
                "tags": [
                    "task"
                ],
                "summary": "Task Create Operation",
                "description": "This is can create a new Task",
                "operationId": "createTask",
                "requestBody": {
                    "description": "Create Task Model Object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {}
                    }
                },
                "x-codegen-request-body-name": "body"
            }
        },
        "/lists/{listId}/tasks/all": {
            "get": {
                "tags": [
                    "task"
                ],
                "summary": "Get all Task",
                "description": "A user can get all Task",
                "operationId": "getAllTask",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/lists/{listId}/tasks/{taskId}": {
            "get": {
                "tags": [
                    "task"
                ],
                "summary": "Find task by id",
                "description": "A user can find one task by her id",
                "operationId": "findOneTask",
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/lists/{listId}/tasks/update/{taskId}": {
            "patch": {
                "tags": [
                    "task"
                ],
                "summary": "Update single task",
                "discription": "a user can update single task",
                "operationId": "updateTask",
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "you can update any of single column data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {}
                }
            }
        },
        "/lists/{listId}/tasks/delete/{taskId}": {
            "delete": {
                "tags": [
                    "task"
                ],
                "summary": "Task delete",
                "discription": "a user can delete spacific task",
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/users/create": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "User Create Operation",
                "description": "this is can create a User",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Create User Model Object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegistrationDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {}
                    }
                },
                "x-codegen-request-body-name": "body"
            }
        },
        "/users/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "User Login Operation",
                "description": "In this section user can login ",
                "operationId": "createLogin",
                "requestBody": {
                    "description": "Login User Model Object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {}
                    }
                },
                "x-codegen-request-body-name": "body"
            }
        },
        "/users/me/access-token": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Find user by id",
                "description": "A user can find one user by her id",
                "operationId": "findUserOne",
                "parameters": [
                    {
                        "name": "_id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "refresh-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserLoginDTO"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserRegistrationDTO": {
                "required": [
                    "name",
                    "phoneNnumber",
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "UserLoginDTO": {
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "ListDTO": {
                "required": [
                    "title",
                    "_userid"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "_userid": {
                        "type": "string"
                    }
                }
            },
            "TaskDTO": {
                "required": [
                    "title",
                    "_listId"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "_listId": {
                        "type": "string"
                    }
                }
            }
        }
    }
}